reorganización de carpetas, renombrado de componentes y mejora de lógica de productos
