+++ plataforma-web-frontend/.env.example
@@ -1,0 +1,15 @@
+# API Configuration
+REACT_APP_API_BASE_URL=http://localhost:8080/api
+REACT_APP_API_TIMEOUT=10000
+
+# Security
+REACT_APP_ENABLE_SECURITY_HEADERS=true
+REACT_APP_CSP_ENABLED=true
+
+# Features
+REACT_APP_ENABLE_ANALYTICS=false
+REACT_APP_ENABLE_ERROR_REPORTING=false
+
+# UI Configuration
+REACT_APP_ITEMS_PER_PAGE=12
+REACT_APP_DEBOUNCE_DELAY=300
