--- plataforma-web-frontend/src/components/preloader/css/Preloader.css
@@ -1,76 +1,0 @@
-/* Fondo cubriendo toda la pantalla */
-.preloader-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background-color: rgba(
-    255,
-    255,
-    255,
-    0.9
-  ); /* Puedes usar rgba(0,0,0,0.7) para fondo oscuro */
-  z-index: 9999;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-/* Contenedor del spinner y el texto */
-.preloader-content {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-/* Spinner de barras */
-.spinner {
-  width: 50px;
-  height: 40px;
-  text-align: center;
-  font-size: 10px;
-  margin-bottom: 12px;
-}
-
-.spinner > div {
-  background-color: #13a7b2; /* Color azul personalizado */
-  height: 100%;
-  width: 6px;
-  display: inline-block;
-  animation: sk-stretchdelay 1.2s infinite ease-in-out;
-  margin: 0 1px;
-}
-
-.spinner .rect2 {
-  animation-delay: -1.1s;
-}
-.spinner .rect3 {
-  animation-delay: -1s;
-}
-.spinner .rect4 {
-  animation-delay: -0.9s;
-}
-.spinner .rect5 {
-  animation-delay: -0.8s;
-}
-
-/* Animación */
-@keyframes sk-stretchdelay {
-  0%,
-  40%,
-  100% {
-    transform: scaleY(0.4);
-  }
-  20% {
-    transform: scaleY(1);
-  }
-}
-
-/* Texto debajo del spinner */
-.preloader-text {
-  font-size: 1.2rem;
-  color: #333;
-  font-weight: 500;
-  text-align: center;
-}
