+++ plataforma-web-frontend/src/features/products/components/ProductCard.jsx
